Fix pybind version conflicts

+++ src/Rtree.cpp
@@ -16,7 +16,7 @@ Rtree::Rtree() {
     this->m_treeSize = 1;
 }
 
-Rtree::Rtree(int m, int M) {
+Rtree::Rtree(size_t m, size_t M) {
     this->m_root = new Node();
     this->m_treeSize = 1;
     this->m_maxChildren = M;
